+++ conf/script/src/build_system/cmd/setup/setup_steps/__init__.py
@@ -1,4 +1,4 @@
-from build_system.cmd.setup.setup_steps.step_create_symlimks_to_compilers_export_shell_env_script import create_symlinks_to_compilers_export_shell_env_script
+from build_system.cmd.setup.setup_steps.step_create_symlimks_to_compilers_export_shell_env_scripts import create_symlinks_to_compilers_export_shell_env_script
 from build_system.cmd.setup.setup_steps.step_create_targets_build_dirs import create_targets_build_dirs
 from build_system.cmd.setup.setup_steps.step_fetch_host_compilers import fetch_host_compilers
 from build_system.cmd.setup.setup_steps.step_get_root_dir import get_verified_root_dir
