+++ conf/script/src/build_system/cmd/setup/setup_steps/step_create_symlimks_to_compilers_export_shell_env_scripts.py
